Enable i18n feature
