+++ example/src/App.js
@@ -1,12 +1,14 @@
-import React, { Fragment } from 'react'
-import WuAuthForm, { ExampleComponent } from 'wu-auth-form-test'
+import React, { useState, Suspense } from 'react'
+import WuAuthForm from 'wu-auth-form-test'
 import 'wu-auth-form-test/dist/index.css'
 
 const App = () => {
+  const [authLayout, setAuthLayout] = useState('sign-up')
+
   return (
-    <>
-      <WuAuthForm></WuAuthForm>
-    </>
+    <Suspense fallback='loading'>
+      <WuAuthForm authLayout={authLayout} setAuthLayout={setAuthLayout} />
+    </Suspense>
   )
 }
