+++ src/components/auth-form/sign-in/sign-in.module.scss
@@ -83,6 +83,11 @@
     margin-right: 10px;
     font-size: 18px;
   }
+
+  .btn-txt {
+    flex-grow: 1;
+    text-align: center;
+  }
 }
 
 .back {
