+++ src/components/auth-form/sign-up/step2.js
@@ -1,24 +1,28 @@
 import classNames from 'classnames/bind'
-import Proptypes from 'prop-types'
 import React, { Fragment } from 'react'
+import Proptypes from 'prop-types'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faArrowLeft } from '@fortawesome/free-solid-svg-icons'
 import { ErrorMessage } from 'formik'
+import { useTranslation } from 'react-i18next'
 
-import styles from './sign-in.module.scss'
+import styles from './sign-up.module.scss'
 
 const cx = classNames.bind(styles)
 
-export default function SignInFormStep2({
+export default function SignUpFormStep2({
   toggleAuthLayout,
   back,
   handleChange,
   handleSubmit,
   isSubmitting,
+  termsOfServiceUrl,
   children
 }) {
+  const { t } = useTranslation()
+
   return (
-    <>
+    <Fragment>
       <FontAwesomeIcon
         icon={faArrowLeft}
         className={cx('back')}
@@ -26,9 +30,27 @@ export default function SignInFormStep2({
       />
       <div className={cx('container--auth-form')}>
         <div className={cx('container--main')}>
+          <div className={cx('title', 'step2')}>
+            {t('signup_form.email_form_title')}
+          </div>
+
           <div className={cx('field')}>
             <input
-              placeholder='帳號- Email'
+              placeholder={t('username.plhdr')}
+              className={cx('text-input')}
+              onChange={handleChange}
+              name='username'
+              type='text'
+            />
+          </div>
+          <ErrorMessage
+            name='username'
+            component='div'
+            className={cx('field__err-msg')}
+          />
+          <div className={cx('field')}>
+            <input
+              placeholder={t('email.plhdr_signup')}
               className={cx('text-input')}
               onChange={handleChange}
               name='email'
@@ -43,54 +65,64 @@ export default function SignInFormStep2({
 
           <div className={cx('field')}>
             <input
-              placeholder='密碼'
+              placeholder={t('pwd.plhdr')}
               className={cx('text-input')}
               onChange={handleChange}
               name='password'
               type='password'
             />
           </div>
+
           <ErrorMessage
             name='password'
             component='div'
             className={cx('field__err-msg')}
           />
 
+          <div className={cx('container--prompt')}>
+            {`${t('terms_of_service.signup_means_agree')}`}
+            <a target='_blank' className={cx('hint')} href={termsOfServiceUrl}>
+              {t('terms_of_service.policies')}
+            </a>
+          </div>
+
           <button
             type='submit'
-            formFor='auth-form'
+            formfor='auth-form'
             className={cx('btn--sign-in')}
             onClick={handleSubmit}
           >
-            登入
+            {t('submit')}
           </button>
 
           <div className={cx('container--prompt')}>
-            還沒有帳號？{' '}
+            {t('has_account')}
             <span
               className={cx('hint')}
               onClick={toggleAuthLayout}
               role='button'
               tabIndex='0'
             >
-              註冊
+              {t('login')}
             </span>
           </div>
           {children}
         </div>
       </div>
-    </>
+    </Fragment>
   )
 }
 
-SignInFormStep2.propTypes = {
-  onFinish: Proptypes.func.isRequired,
-  onRequest: Proptypes.func.isRequired,
+SignUpFormStep2.propTypes = {
+  onFinish: Proptypes.func,
+  onRequest: Proptypes.func,
   onSuccess: Proptypes.func,
   onFail: Proptypes.func
 }
 
-SignInFormStep2.defaultProps = {
+SignUpFormStep2.defaultProps = {
+  onFinish: null,
+  onRequest: null,
   onSuccess: null,
   onFail: null
 }
