+++ src/index.js
@@ -1,4 +1,6 @@
 import React from 'react'
-import AuthForm from './components/auth-form-v2/index'
+
+import './i18n'
+import AuthForm from './components/auth-form/index'
 
 export default AuthForm
