+++ src/lib/validator/auth-fields.js
@@ -1,48 +1,46 @@
-import _isEmpty from 'lodash.isempty';
+import _isEmpty from 'lodash.isempty'
 
-export const isValidEmail = email =>
-  email && /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(email);
+import i18n from '../../i18n'
 
-export const isValidPwd = password => password && password.length >= 6;
+export const isValidEmail = (email) =>
+  email && /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(email)
+
+export const isValidPwd = (password) => password && password.length >= 6
 
 export function signUpValidate(values) {
-  const errors = {};
-  const { username, email, password } = values;
+  const errors = {}
+  const { username, email, password } = values
 
   if (_isEmpty(username)) {
-    errors.username = '必填欄位';
+    errors.username = i18n.t('required_field')
   }
 
-  if (_isEmpty(password)) {
-    errors.password = '必填欄位';
-  } else if (!isValidPwd(password)) {
-    errors.password = '密碼至少為 6 位英數字';
-  }
+  errors.password = validatePwd(password)
+  errors.email = validateEmail(email)
 
-  if (_isEmpty(email)) {
-    errors.email = '必填欄位';
-  } else if (!isValidEmail(email)) {
-    errors.email = 'Email 格式錯誤';
-  }
-
-  return errors;
+  return errors
 }
 
 export function signInValidate(values) {
-  const errors = {};
-  const { email, password } = values;
+  const errors = {}
+  const { email, password } = values
 
-  if (_isEmpty(password)) {
-    errors.password = '必填欄位';
-  } else if (!isValidPwd(password)) {
-    errors.password = '密碼至少為 6 位英數字';
-  }
+  errors.password = validatePwd(password)
+  errors.email = validateEmail(email)
 
-  if (_isEmpty(email)) {
-    errors.email = '必填欄位';
-  } else if (!isValidEmail(email)) {
-    errors.email = 'Email 格式錯誤';
-  }
+  return errors
+}
 
-  return errors;
+function validatePwd(password) {
+  if (_isEmpty(password)) return i18n.t('required_field')
+  if (!isValidPwd(password)) return i18n.t('pwd.plhdr_set')
+
+  return null
+}
+
+function validateEmail(email) {
+  if (_isEmpty(email)) return i18n.t('required_field')
+  if (!isValidEmail(email)) return i18n.t('email.err')
+
+  return null
 }
